+++ test/geiger_toolbox_apis_test.dart
@@ -2,11 +2,6 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:geiger_toolbox_apis/geiger_toolbox_apis.dart';
 
-void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
-  });
+void main() async {
+  await GeigerToolboxAPIs.connect();
 }