+++ CHANGELOG.md
@@ -1,3 +1,20 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Init the package
+
+## 0.0.2
+
+* Fix the problem of missing path package
+* Made all the APIs are static
+
+## 0.0.3
+* fix getSensingData value query
+
+## 0.0.4
+* fix getSensingData value query - with raw query
+
+## 0.0.5
+* fix getSensingData value query - with raw query - return data
+
+## 0.0.6
+* fix getSensingData value query